+++ docs/util/get_figure_examples.sh
@@ -13,13 +13,27 @@ figshare_version=v4
 tools_manager=singularity
 
 species=worm ; cd $test_dir/$tools_manager/$species
-nextflow run  jsalignon/cactus -r hotfix/0.8.1  -profile $tools_manager --executor_local_cpus $cpu_nb --executor_local_memory $memory_size --references_dir $cactus_dir/references/$figshare_version -params-file parameters/full_test.yml --split__threshold_values [1000] --res_dir 'results/almost_full_test'
+# nextflow run  jsalignon/cactus -r hotfix/0.8.1  -profile $tools_manager --executor_local_cpus $cpu_nb --executor_local_memory $memory_size --references_dir $cactus_dir/references/$figshare_version -params-file parameters/full_test.yml --split__threshold_values [1000] --res_dir 'results/almost_full_test'
+nextflow run  jsalignon/cactus -r v0.9.0  -profile $tools_manager \
+  --executor_local_cpus $cpu_nb --executor_local_memory $memory_size \
+  --references_dir $cactus_dir/references/$figshare_version \
+  -params-file parameters/full_test.yml --split__threshold_values [1000] \
+  --res_dir 'results/v0.9.0'
+
+nextflow run  jsalignon/cactus -r revisions  -profile $tools_manager \
+  --executor_local_cpus $cpu_nb --executor_local_memory $memory_size \
+  --references_dir $cactus_dir/references/$figshare_version \
+  -params-file parameters/full_test.yml --split__threshold_values [1000] \
+  --res_dir 'results/v0.9.0_revisions'
+
 # nextflow run  jsalignon/cactus -r main -latest -profile $tools_manager --executor_local_cpus $cpu_nb --executor_local_memory $memory_size --references_dir $cactus_dir/references/$figshare_version -params-file parameters/full_test.yml --split__threshold_values [1000] --res_dir 'results/almost_full_test'
 # nextflow run  jsalignon/cactus -r main -latest -profile $tools_manager --executor_local_cpus $cpu_nb --executor_local_memory $memory_size --references_dir $cactus_dir/references/$figshare_version -params-file parameters/no_enrich__no_rtr.yml
 
 # species=fly ; cd $test_dir/$tools_manager/$species
 # nextflow run  jsalignon/cactus -r main -latest -profile $tools_manager --executor_local_cpus $cpu_nb --executor_local_memory $memory_size --references_dir $cactus_dir/references/$figshare_version -params-file parameters/no_enrich.yml
 # nextflow run  jsalignon/cactus -r main -latest -profile $tools_manager --executor_local_cpus $cpu_nb --executor_local_memory $memory_size --references_dir $cactus_dir/references/$figshare_version -params-file parameters/no_enrich__no_gtr.yml
+
+
 
 
 
@@ -43,7 +57,8 @@ test_dir_2="$cactus_dir/$testing_dir_name/singularity"
 res_worm_dir="$test_dir_2/worm/results/"
 res_fly_dir="$test_dir_2/fly/results/"
 
-run_worm_dir="$res_worm_dir/almost_full_test"
+# run_worm_dir="$res_worm_dir/almost_full_test"
+run_worm_dir="$res_worm_dir/v0.9.0_revisions"
 run_worm_figure_dir="$run_worm_dir/Figures_Individual"
 run_worm_tables_dir="$run_worm_dir/Tables_Merged"
 
@@ -77,15 +92,17 @@ find $run_worm_figure_dir -name "spearman_correlation_heatmap_without_outliers_w
 find $run_worm_figure_dir -name "pca_top5000_without_control_pca.pdf" -exec cp "{}" $examples_dir_pdf \;
 find $run_worm_figure_dir -regex ".*\(ATAC\|mRNA\)__multiQC.html" -exec cp "{}" $examples_dir_html \;
 
-
-
+# Panels for tutorial
+find $run_worm_figure_dir -regex ".*ctl__genes_self__heatmap.pdf" -exec cp "{}" $examples_dir_pdf \;
+find $run_worm_figure_dir -regex ".*ctl__func_anno_KEGG__heatmap.pdf" -exec cp "{}" $examples_dir_pdf \;
+find $run_worm_figure_dir -regex ".*hmg4_3__saturation.*.pdf" -exec cp "{}" $examples_dir_pdf \;
 
 ## Making png images from PDF
 cd $examples_dir_pdf
 
 # for FILE in $(ls ATAC__all__down__200__hmg4_vs_spt16__motifs__barplot.pdf) 
 # for FILE in $(ls mRNA_volcano*.pdf) 
-for FILE in $(ls *heatmap*.pdf) 
+for FILE in $(ls *hmg4_3__saturation*.pdf) 
 # for FILE in $(ls hmg4_vs_spt16__ATAC_volcano.pdf) 
 # for FILE in $(ls *.pdf) 
 do